requirements.txt
python-telegram-bot==20.0
python-dotenv==0.19.2
openai
binance
pillow
pydub
ffmpeg-python
pytz
apscheduler
nest_asyncio
pytesseract
